speed up startup by using org tangle only when config.org changed
To determine if config is changed, we can calculate its current md5 and compare it with previous
one. If file with previous checksum doesn't exist, then we should bootstrap config anyway, so
this will invoke tangle. In most cases, Emacs starts with the same config, so this is a nice
optimization.